fix column data type

+++ data/create-tables.sql
@@ -3,14 +3,11 @@ CREATE TABLE taxis(
    plate VARCHAR(9) NOT NULL
 );
 
--- INSERT INTO trajectories (taxi_id, date, latitude, longitude )
--- VALUES ('10133','2008-02-02 16:18:49',116.34692,39.93064);
-
 CREATE TABLE trajectories (
     id INT GENERATED ALWAYS AS IDENTITY,
     taxi_id INT NOT NULL,
     date TIMESTAMP NOT NULL,
-    latitude BIGINT NOT NULL,
-    longitude BIGINT NOT NULL,
+    latitude DOUBLE PRECISION NOT NULL,
+    longitude DOUBLE PRECISION NOT NULL,
     CONSTRAINT fk_taxi FOREIGN KEY(taxi_id) REFERENCES taxis(id)
 );
